[TASK] Filter trailing number signs (dot and comma) from localized input
When user enters numbers with trailing dot/comma at the moment those
are recognized by LocalizedToNormalized filter as "date" because
Zend_Locale_Format::checkDateFormat returns true on this value.
This leads to an array with date information being returned which cannot
be processed subsequently.

By just removing training dots and commas we mitigate that problem.